Lab6_Statful_Systems/requirements.txt
pytest
grpcio
grpcio-tools